requirements-dev.txt
appdirs
attrs
sqlitedict
#
# dev
ismain
wheel
black
pytest
pytest-cov
pytest-pycharm
flake8
twine
mypy
appdirs-stubs
